create team class
