+++ lib/classes/team.js
@@ -1,0 +1,9 @@
+class Team {
+	constructor(manager, engineers, interns) {
+		this.manager = manager;
+		this.engineers = engineers;
+		this.interns = interns;
+	}
+}
+
+module.exports = Team
