+++ lib/functions/buildTeam.js
@@ -1,16 +1,9 @@
 const inquirer = require('inquirer');
-const { createManager } = require('./inquirer-prompts/createManager.js');
-const { createEngineer } = require('./inquirer-prompts/createEngineer.js');
-const { createIntern } = require('./inquirer-prompts/createIntern.js');
-const { getNextStep } = require('./inquirer-prompts/getNextStep.js');
-
-class Team {
-	constructor(manager, engineers, interns) {
-		this.manager = manager;
-		this.engineers = engineers;
-		this.interns = interns;
-	}
-}
+const { createManager } = require('../inquirer-prompts/createManager.js');
+const { createEngineer } = require('../inquirer-prompts/createEngineer.js');
+const { createIntern } = require('../inquirer-prompts/createIntern.js');
+const { getNextStep } = require('../inquirer-prompts/getNextStep.js');
+const { Team } = require('../classes/Team.js');
 
 let team = new Team();
